Change the LastUpdated/Published date to derive from the Created date
In most cases, I'm fairly certain the package will never be modified, only ever created.
It's possible that the Nuget PhysicalFileSystem lib should be changed to fix this and return the
created date if the file has never been modified. Not sure how it would detect this, but we can pass
it up the chain if that makes more sense.